pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "hypersync"
requires-python = ">=3.9"
dependencies = ["strenum>=0.4.15,<0.4.16", "python-dotenv>=1.0.0"]
classifiers = [
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Development Status :: 3 - Alpha",
  "Topic :: Database :: Front-Ends",
  "Topic :: Software Development :: Libraries",
]
dynamic = ["version"]

[tool.maturin]
features = ["pyo3/extension-module"]

# See https://www.maturin.rs/metadata.html for options
[project.urls]
homepage = "https://envio.dev/"
documentation = "https://docs.envio.dev/docs/HyperSync/overview"
repository = "https://github.com/enviodev/hypersync-client-python"
